Relation-User&BlogEstablished

+++ backend/src/handlers/handlers.ts
@@ -5,7 +5,7 @@ import User from "../models/User"
 import Blog from "../models/Blog"
 
 import Comment from "../models/Comments"
-import {Document} from "mongoose"
+import {Document, startSession} from "mongoose"
 import {compareSync, hashSync} from "bcryptjs"
 
 const RootQuery=new GraphQLObjectType({
@@ -94,14 +94,29 @@ const mutations=new GraphQLObjectType({
                 title:{type:new GraphQLNonNull(GraphQLString)},
                 content:{type:new GraphQLNonNull(GraphQLString)},  
                 date:{type:new GraphQLNonNull(GraphQLString)},
+                user:{type:new GraphQLNonNull(GraphQLID)},
             },
             async resolve(parent,args){
                 let blog:Document<any,any,any>
+                //to populate the data to refrenced collection we need to create a session
+                const session =await startSession();
                try {
-                blog=new Blog({title:args.title,content:args.content,date:args.date});
-               return await blog.save();
+                session.startTransaction({session});
+                blog=new Blog({title:args.title,content:args.content,date:args.date,user:args.user});   
+                //checking user if we have the user with parent user id(line 97)
+                const existingUser = await User.findById(args.user)
+                if(!existingUser) return new Error("User Not Found! terminating session");
+                //after creating instance of blog the we need to start the 
+                //transaction
+            
+                //then we need to commit the transaction whic comes under finaly, before commiting we need to save the record in both the user and the blog.
+                existingUser.blogs.push(blog);
+                await existingUser.save({session});
+               return await blog.save({session});
                } catch (error) {
                 return new Error(error.message);
+               }finally{
+                await session.commitTransaction();
                }
             }
         },
@@ -136,16 +151,28 @@ const mutations=new GraphQLObjectType({
             },
             async resolve(parent,{id}){
                 let existingBlog:Document<any,any,any>
+                //in delte we have to delte all the comments of user and the blog of user 
+                const session=await startSession();
                 try {
-                    existingBlog=await Blog.findById(id);
+                   session.startTransaction({session})
+                    existingBlog=await Blog.findById(id).populate("user");
+                   //@ts-ignore
+                    const existingUser=existingBlog?.user
+                    if(!existingUser)return new Error(`Could not find linked user to this blog!terminating the session`)
+
                     if(!existingBlog) return new Error("No Blog found");
+                    //pull method in mongoose will pull out the record 
+                    existingUser.blogs.pull(existingBlog);
+                    await existingUser.save({session});
                     return await Blog.findByIdAndRemove(id);
                 } catch (error) {
                     return new Error(error.message);
+                }finally{
+                    session.commitTransaction();
                 }
             }
         }
-
+        
     }
 })
 //now we have to make queries inside the graphqlschema
